Tree: make public APIs alpha
