+++ packages/dds/tree/api-extractor.json
@@ -1,4 +1,20 @@
 {
 	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-	"extends": "@fluidframework/build-common/api-extractor-common-strict.json"
+	"extends": "@fluidframework/build-common/api-extractor-common-strict.json",
+	/**
+	 * Configures how the .d.ts rollup file will be generated.
+	 */
+	"dtsRollup": {
+		"enabled": true,
+		"alphaTrimmedFilePath": "<projectFolder>/dist/<unscopedPackageName>.d.ts"
+	},
+	"messages": {
+		"extractorMessageReporting": {
+			// A type signature should not reference another types whose release tag is less visible.
+			"ae-missing-release-tag": {
+				"logLevel": "error",
+				"addToApiReportFile": false
+			}
+		}
+	}
 }
