+++ packages/dds/tree/src/core/dependency-tracking/incrementalObservation.ts
@@ -11,6 +11,7 @@ import { Dependent, Dependee } from "./dependencies";
 
 /**
  * State/Environment for a computation which can observe {@link Dependees}.
+ * @alpha
  */
 export interface ObservingContext {
 	/**
@@ -28,6 +29,7 @@ export interface ObservingContext {
 /**
  * Ensures this context's computation is a dependant to dependee (adding it if needed).
  * Ensures this context's computation has dependee as a dependee (adding it if needed).
+ * @alpha
  */
 export function recordDependency(
 	dependent: ObservingDependent | undefined,
@@ -42,6 +44,7 @@ export function recordDependency(
 
 /**
  * A dependent which can have observations it makes recorded onto it.
+ * @alpha
  */
 export interface ObservingDependent extends Dependent {
 	/**
