+++ packages/dds/tree/src/core/tree/mapTree.ts
@@ -12,7 +12,7 @@ import { FieldKey, NodeData } from "./types";
 /**
  * Simple in memory tree representation based on Maps.
  * MapTrees should not store empty fields.
- * @public
+ * @alpha
  */
 export interface MapTree extends NodeData {
 	fields: Map<FieldKey, MapTree[]>;
