+++ packages/dds/tree/src/core/tree/types.ts
@@ -15,6 +15,7 @@ import { GlobalFieldKeySymbol, symbolFromKey } from "./globalFieldKeySymbol";
  * To avoid collisions, we can not abstract over local and global field keys using the same format for each
  * (that would make telling them apart impossible).
  * Thus global field keys are using their symbols instead.
+ * @alpha
  */
 export type FieldKey = LocalFieldKey | GlobalFieldKeySymbol;
 
@@ -22,6 +23,9 @@ export function isLocalKey(key: FieldKey): key is LocalFieldKey {
 	return typeof key === "string";
 }
 
+/**
+ * @alpha
+ */
 export type TreeType = TreeSchemaIdentifier;
 
 /**
@@ -35,6 +39,7 @@ export type TreeType = TreeSchemaIdentifier;
  * This has to be a LocalFieldKey since different nodes will have different FieldSchema for it.
  * This makes it prone to collisions and suggests
  * that this intention may be better conveyed by metadata on the TreeViewSchema.
+ * @alpha
  */
 export const EmptyKey: LocalFieldKey = brand("");
 
@@ -42,6 +47,7 @@ export const EmptyKey: LocalFieldKey = brand("");
  * GlobalFieldKey to use for the root of documents.
  * TODO: if we do want to standardize on a single value for this,
  * it likely should be namespaced or a UUID to avoid risk of collisions.
+ * @alpha
  */
 export const rootFieldKey: GlobalFieldKey = brand("rootFieldKey");
 export const rootFieldKeySymbol: GlobalFieldKeySymbol = symbolFromKey(rootFieldKey);
@@ -50,7 +56,7 @@ export const rootField = keyAsDetachedField(rootFieldKeySymbol);
 /**
  * Location of a tree relative to is parent container (which can be a tree or forest).
  *
- * @public
+ * @alpha
  */
 export interface ChildLocation {
 	readonly container: ChildCollection;
@@ -59,6 +65,7 @@ export interface ChildLocation {
 
 /**
  * Wrapper around DetachedField that can be detected at runtime.
+ * @alpha
  */
 export interface RootField {
 	readonly key: DetachedField;
@@ -66,6 +73,7 @@ export interface RootField {
 
 /**
  * Identifier for a child collection, either on a node/tree or at the root of a forest.
+ * @alpha
  */
 export type ChildCollection = FieldKey | RootField;
 
@@ -84,6 +92,7 @@ export type ChildCollection = FieldKey | RootField;
  *
  * In some APIs DetachedFields are used as LocalFieldKeys on a special implicit root node
  * to simplify the APIs and implementation.
+ * @alpha
  */
 export interface DetachedField extends Opaque<Brand<string, "tree.DetachedField">> {}
 
@@ -92,6 +101,7 @@ export interface DetachedField extends Opaque<Brand<string, "tree.DetachedField"
  * This maps detached field to field keys for thus use.
  *
  * @returns `field` as a {@link FieldKey} usable on a special root node serving as a parent of detached fields.
+ * @alpha
  */
 export function detachedFieldAsKey(field: DetachedField): FieldKey {
 	if (field === rootField) {
@@ -104,6 +114,7 @@ export function detachedFieldAsKey(field: DetachedField): FieldKey {
  * The inverse of {@link detachedFieldAsKey}.
  * Thus must only be used on {@link LocalFieldKey}s which were produced via {@link detachedFieldAsKey},
  * and with the same scope (ex: forest) as the detachedFieldAsKey was originally from.
+ * @alpha
  */
 export function keyAsDetachedField(key: FieldKey): DetachedField {
 	if (isLocalKey(key)) {
@@ -124,6 +135,7 @@ export function keyAsDetachedField(key: FieldKey): DetachedField {
  * TODO: integrate this into Schema. Decide how to persist them (need stable Id?). Maybe allow updating field kinds?.
  * TODO: make families of changes per field kind. Build editing APIs from that.
  * TODO: factor ChangeRebaser implementations to support adding new field kinds.
+ * @alpha
  */
 export interface FieldKind {
 	readonly name: string;
@@ -141,17 +153,19 @@ export interface FieldKind {
  * Use this type instead of directly using Serializable for both clarity and so the above TODO can be addressed.
  *
  * This is a named interface instead of a Type alias so tooling (ex: refactors) will not replace it with `any`.
+ * @alpha
  */
 export interface TreeValue extends Serializable {}
 
 /**
  * Value stored on a node.
+ * @alpha
  */
 export type Value = undefined | TreeValue;
 
 /**
  * The fields required by a node in a tree.
- * @public
+ * @alpha
  */
 export interface NodeData {
 	/**
