+++ packages/dds/tree/src/feature-libraries/defaultSchema.ts
@@ -14,12 +14,14 @@ export const neverField = fieldSchema(value, []);
 
 /**
  * FieldSchema which is impossible to put anything in.
+ * @alpha
  */
 
 export const emptyField = fieldSchema(forbidden, []);
 
 /**
  * TreeSchema which is impossible for any data to be in schema with.
+ * @alpha
  */
 export const neverTree: TreeSchema = {
 	localFields: emptyMap,
@@ -34,6 +36,7 @@ export const neverTree: TreeSchema = {
  *
  * This requires new node types to have explicit stored schema to exist in documents,
  * and allows adding new global fields along with their schema at any point.
+ * @alpha
  */
 export const defaultSchemaPolicy: FullSchemaPolicy = {
 	fieldKinds,
