+++ packages/dds/tree/src/feature-libraries/modular-schema/modularChangeFamily.ts
@@ -43,6 +43,7 @@ import { decodeJsonFormat0, encodeForJsonFormat0 } from "./modularChangeEncoding
  * as determined by the schema.
  *
  * @sealed
+ * @alpha
  */
 export class ModularChangeFamily
 	implements ChangeFamily<ModularEditBuilder, ModularChangeset>, ChangeRebaser<ModularChangeset>
@@ -453,6 +454,7 @@ class ModularChangeEncoder extends ChangeEncoder<ModularChangeset> {
 
 /**
  * @sealed
+ * @alpha
  */
 export class ModularEditBuilder
 	extends ProgressiveEditBuilderBase<ModularChangeset>
