+++ packages/dds/tree/src/util/brand.ts
@@ -15,6 +15,7 @@ import { Invariant, isAny } from "./typeCheck";
  * `Type 'Name1' is not assignable to type 'Name2'.`
  *
  * These branded types are not opaque: A `Brand<A, B>` can still be used as a `B`.
+ * @alpha
  */
 export type Brand<ValueType, Name extends string> = ValueType & BrandedType<ValueType, Name>;
 
@@ -32,6 +33,7 @@ export type Brand<ValueType, Name extends string> = ValueType & BrandedType<Valu
  * but the compiler may think its true in some cases.
  *
  * @sealed
+ * @alpha
  */
 export abstract class BrandedType<ValueType, Name extends string> {
 	protected _typeCheck?: Invariant<ValueType>;
@@ -61,6 +63,7 @@ export abstract class BrandedType<ValueType, Name extends string> {
  * ```typescript
  * export interface MyType extends Opaque<Brand<string, "myPackage.MyType">>{}
  * ```
+ * @alpha
  */
 export type Opaque<T extends Brand<any, string>> = T extends Brand<infer ValueType, infer Name>
 	? BrandedType<ValueType, Name>
@@ -68,6 +71,7 @@ export type Opaque<T extends Brand<any, string>> = T extends Brand<infer ValueTy
 
 /**
  * See {@link extractFromOpaque}.
+ * @alpha
  */
 export type ExtractFromOpaque<TOpaque extends BrandedType<any, string>> =
 	TOpaque extends BrandedType<infer ValueType, infer Name>
@@ -79,6 +83,7 @@ export type ExtractFromOpaque<TOpaque extends BrandedType<any, string>> =
 /**
  * Implementation detail of type branding. Should not be used directly outside this file,
  * but shows up as part of branded types so API-Extractor requires it to be exported.
+ * @alpha
  */
 export type ValueFromBranded<T extends BrandedType<any, string>> = T extends BrandedType<
 	infer ValueType,
@@ -90,6 +95,7 @@ export type ValueFromBranded<T extends BrandedType<any, string>> = T extends Bra
 /**
  * Implementation detail of type branding. Should not be used directly outside this file,
  * but shows up as part of branded types so API-Extractor requires it to be exported.
+ * @alpha
  */
 export type NameFromBranded<T extends BrandedType<any, string>> = T extends BrandedType<
 	any,
@@ -103,6 +109,7 @@ export type NameFromBranded<T extends BrandedType<any, string>> = T extends Bran
  *
  * It is assumed that only code that produces these "opaque" handles does this conversion,
  * allowing these handles to be considered opaque.
+ * @alpha
  */
 export function extractFromOpaque<TOpaque extends BrandedType<any, string>>(
 	value: TOpaque,
@@ -114,6 +121,7 @@ export function extractFromOpaque<TOpaque extends BrandedType<any, string>>(
  * Adds a type {@link Brand} to a value.
  *
  * Only do this when specifically allowed by the requirements of the type being converted to.
+ * @alpha
  */
 export function brand<T extends Brand<any, string>>(
 	value: T extends BrandedType<infer ValueType, string> ? ValueType : never,
@@ -125,6 +133,7 @@ export function brand<T extends Brand<any, string>>(
  * Adds a type {@link Brand} to a value, returning it as a  {@link Opaque} handle.
  *
  * Only do this when specifically allowed by the requirements of the type being converted to.
+ * @alpha
  */
 export function brandOpaque<T extends BrandedType<any, string>>(
 	value: isAny<ValueFromBranded<T>> extends true ? never : ValueFromBranded<T>,
